fix(audio): タグ情報の取得する際、nullチェックをしていなかったため修正しました。

+++ source/AudioResource.cpp
@@ -6,24 +6,27 @@ namespace Win32
 {
 	AudioMetaData2::AudioMetaData2(const TagLib::Tag* tag, bool disableImage)
 	{
-		allTag =
-			tag->title().toWString() + SEPARATOR +
-			tag->artist().toWString() + SEPARATOR +
-			tag->album().toWString() + SEPARATOR +
-			tag->genre().toWString() + SEPARATOR +
-			tag->comment().toWString();
-
-		t = tag->track();
-		y = tag->year();
-//		if (!disableImage)
-		if (false)
+		if (tag)
 		{
-			std::ofstream outFile{ "cover.jpeg", std::ios::out | std::ios::binary };
-			auto pictureData = tag->complexProperties("PICTURE").front()["data"].toByteVector();
-			outFile.write(pictureData.data(), pictureData.size());
-			outFile.close();
-			j = Image{ L"cover.jpeg" };
-			std::filesystem::remove(L"cover.jpeg");
+			allTag =
+				tag->title().toWString() + SEPARATOR +
+				tag->artist().toWString() + SEPARATOR +
+				tag->album().toWString() + SEPARATOR +
+				tag->genre().toWString() + SEPARATOR +
+				tag->comment().toWString();
+
+			t = tag->track();
+			y = tag->year();
+	//		if (!disableImage)
+			if (false)
+			{
+				std::ofstream outFile{ "cover.jpeg", std::ios::out | std::ios::binary };
+				auto pictureData = tag->complexProperties("PICTURE").front()["data"].toByteVector();
+				outFile.write(pictureData.data(), pictureData.size());
+				outFile.close();
+				j = Image{ L"cover.jpeg" };
+				std::filesystem::remove(L"cover.jpeg");
+			}
 		}
 	}
